+++ tests/c_module_a/shared.zlcm.h
@@ -1,5 +1,0 @@
-ZLC_MODULE(shared)
-ZLC_HEADER("shared.h")
-ZLC_SOURCE("shared.c")
-
-ZLC_FN(increment, zlcm_shared_a_increment, int, ZLC_PARAM(value, int))