+++ tests/c_module_b/shared.zlcm.h
@@ -1,5 +1,0 @@
-ZLC_MODULE(shared)
-ZLC_HEADER("shared.h")
-ZLC_SOURCE("shared.c")
-
-ZLC_FN(twice, zlcm_shared_b_twice, int, ZLC_PARAM(value, int))