+++ tests/c_module_errors/empty_struct.zlcm.h
@@ -1,3 +1,0 @@
-ZLC_MODULE(empty_struct)
-ZLC_STRUCT(EmptyStruct)
-ZLC_FN(noop, zlcm_empty_noop, int)