+++ tests/c_module_errors/lowercase_struct.zlcm.h
@@ -1,3 +1,0 @@
-ZLC_MODULE(lowercase_struct)
-ZLC_STRUCT(point, ZLC_FIELD(x, float))
-ZLC_FN(noop, zlcm_lowercase_noop, int)