+++ tests/c_module_errors/recursive_struct.zlcm.h
@@ -1,3 +1,0 @@
-ZLC_MODULE(recursive_struct)
-ZLC_STRUCT(Node, ZLC_FIELD(next, Node))
-ZLC_FN(noop, zlcm_recursive_noop, int)