+++ tests/c_module_errors/unsupported_abi_type.zlcm.h
@@ -1,2 +1,0 @@
-ZLC_MODULE(unsupported_abi_type)
-ZLC_FN(noop, zlcm_unsupported_noop, matrix)