+++ tests/native_math.c
@@ -1,5 +1,0 @@
-#include "native_math.h"
-
-int zlcm_native_add(int a, int b) {
-    return a + b;
-}